refactor(spiral): move render+inks into internal/spiral
